08/17/07 - Fixed room editing permissions. 08/17/07 - Added additional bug handling, and protected room 0. Upgrade to version .96 08/18/07 - Begun addition of synchronization to prevent concurrent modification errors. 08/18/07 - Added Death and Kill Count 08/18/07 - Added uptime command. 08/18/07 - Fixed spellcasting maxpos Upgrade to version .97

+++ TrollAttack/Background.java
@@ -13,7 +13,7 @@ package TrollAttack;
  * Preferences - Java - Code Style - Code Templates
  */
 public class Background extends Thread {
-    int time = 0;
+    long time = 0;
 
     // Order by most frequent to least frequent for organization.
     public void run() {
@@ -56,7 +56,7 @@ public class Background extends Thread {
         TrollAttack.message("Background shutting down.");
     }
 
-    public int getTime() {
+    public long getTime() {
         return time;
 
     }
